HistoricProcess/requirements.txt
# Function dependencies, for example:
# package>=version
pandas==1.4.4
requests==2.28.1
google.cloud.storage==2.5.0
google.cloud.bigquery==3.3.2
arcgis==1.9.0
datalab==1.2.0
pyarrow==6.0.1
gcsfs==2021.11.1
numpy==1.19.5
db-dtypes==1.0.3